move newshape & activeshape attributes & defaults to components/shapes/draw_newshape
